requirements.txt
<<<<<<< HEAD
Flask==2.3.2
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.0
=======
Flask==2.3.2
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.0
gunicorn==20.1.0
>>>>>>> master
